Adds black and ruff pre-commit hook and reformats code
This adds a pre-commit hook, that requires each user to install
so that when they run `git commit` it will run right before that
and adjust/complain about code not looking standardized.

As you can see the code has been reformatted.

See https://black.readthedocs.io/en/stable/ and
https://beta.ruff.rs/docs/ for information on black and ruff.
Ruff takes the place of flake8 if you're familiar with that.